requirement.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langgraph>=0.0.1
langchain_openai>=0.1.0
langchain_chroma>=0.1.0
langchain_text_splitters>=0.1.0
requests>=2.31.0
bs4>=0.0.2
beautifulsoup4>=4.12.0
chromadb>=1.4.0
openai>=1.3.0
python-dotenv>=1.0.0
pypdf>=3.17.0
langchain_ollama